D: 第一版序列化重构， 将ResourceManager写为 Loader形式，二进制序列化加载
将ResrouceManager重构， 用二进制序列化进行 同步异步加载。

+++ Runtime/Resources/Asset.h
@@ -12,7 +12,7 @@ namespace EngineCore
     {
         uint64_t value = 0;
         explicit operator bool() const {return value != 0;};
-        operator uint32_t() const { return value;}
+        operator uint64_t() const { return value;}
         bool IsValid() const {return value != 0;};
         inline void Reset() { value = 0; };
         AssetID() = default;
